server/requirements.txt
# Eigensparse - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
